02-crzp-agent/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
pytest>=7.4.0
hypothesis>=6.82.0
requests>=2.31.0
playwright>=1.44.0
asyncio

